apps/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-cloud-vision==3.4.4
python-dotenv==1.0.0 